# Fix garbled URL preview for non UTF-8 HTML, with optimization