@vitorpamplona 2024-03-26
Any chance you can add your benchmark to the project as well? We can move the parser to `commons` to give access to the benchmark module